requirements.txt
#pip freeze > requirements.txt
asgiref==3.2.10
astroid==2.4.2
#bootstrap-admin==0.4.4
cffi==1.14.1
cryptography==3.0
dj-database-url==0.5.0
Django==3.0.6
django-environ==0.4.5
#django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dry-rest-permissions==0.1.10
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
